refactor: Nest virtual-tag-configs tools.
