+++ src/tools/index.ts
@@ -11,7 +11,6 @@ import "./cost-providers";
 import "./cost-reports";
 import "./cost-services";
 import "./costs";
-import "./create-virtual-tag-config";
 import "./current-user";
 import "./dashboards";
 import "./financial-commitment-reports";
@@ -29,4 +28,5 @@ import "./tags";
 import "./teams";
 import "./user-feedback";
 import "./users";
+import "./virtual-tag-configs";
 import "./workspaces";
