+++ src/tools/virtual-tag-configs/create-virtual-tag-config.ts
@@ -1,7 +1,7 @@
 import z from "zod";
-import dateValidator from "../utils/dateValidator";
-import MCPUserError from "./structure/MCPUserError";
-import registerTool from "./structure/registerTool";
+import dateValidator from "../../utils/dateValidator";
+import MCPUserError from "../structure/MCPUserError";
+import registerTool from "../structure/registerTool";
 
 const description = `
 Create a Virtual Tag Config in Vantage.
