+++ src/tools/virtual-tag-configs/index.ts
@@ -1,0 +1,1 @@
+import "./create-virtual-tag-config";
